tracker/requirements.txt
opencv-contrib-python
Pillow
scipy
tensorflow